Building and testing a UBI-based image (#382)
This PR adds a UBI-based vault-csi-provider image to the Dockerfile
and CI build workflow. This will produce a UBI image of
vault-csi-provider for arm64 and amd64 architectures, which will be
pushed to docker and redhat container registries at release time.

The bats integration tests and Makefile setup were also modified to
run the tests against the UBI image on both kind and
OpenShift. Instructions for running tests on OpenShift are in the
README, and the CI tests workflow smoke tests the UBI image with the
latest vault and k8s combo.